+++ workflows/determine_mutations.nf
@@ -1,11 +1,13 @@
 include { nextclade } from './process/nextclade'
+include { add_aainsertions } from '../modules/add_aainsertions.nf'
 
 workflow determine_mutations_wf {
     take: 
         fasta  
     main:
         nextclade(fasta)
+        add_aainsertions(nextclade.out)
 
     emit:
-        nextclade.out
+        add_aainsertions.out
 }
